# crossbeam-channel: Remove dependency on smallvec