This doesn't seem to affect benchmarks.