@stjepang 2019-04-03
On my machine, `cargo run --release --bin crossbeam-channel` yields this (showing only the `*select*` tests):

```
bounded_select_both       Rust crossbeam-channel   0.479 sec
bounded_select_rx         Rust crossbeam-channel   0.486 sec
unbounded_select_both     Rust crossbeam-channel   0.432 sec
unbounded_select_rx       Rust crossbeam-channel   0.461 sec
```

After replacing `SmallVec` with `Vec` I get:

```
bounded_select_both       Rust crossbeam-channel   0.587 sec
bounded_select_rx         Rust crossbeam-channel   0.581 sec
unbounded_select_both     Rust crossbeam-channel   0.545 sec
unbounded_select_rx       Rust crossbeam-channel   0.550 sec
```

There seems to be a consistent slowdown of 15-20%.

What benchmarks did you try?