@tuxether 2019-04-03
Here's a sample run from my machine:

Before:
```
bounded_select_both       Rust crossbeam-channel   0.386 sec
bounded_select_rx         Rust crossbeam-channel   0.475 sec
unbounded_select_both     Rust crossbeam-channel   0.403 sec
unbounded_select_rx       Rust crossbeam-channel   0.432 sec
```

After:
```
bounded_select_both       Rust crossbeam-channel   0.384 sec
bounded_select_rx         Rust crossbeam-channel   0.440 sec
unbounded_select_both     Rust crossbeam-channel   0.414 sec
unbounded_select_rx       Rust crossbeam-channel   0.398 sec
```

I'm using glibc 2.28 for the malloc implementation. I should mention that I have noticed regressions when using musl, but that has to do with the fact that the malloc implementation in musl is very primitive.

Feel free to close the PR if there are machines on which this does regress the benchmarks.