@stjepang 2019-04-03
Interesting. Could you perhaps also say what is your machine (CPU) and rustc version?