@BurntSushi 2019-04-03
Here are the benchmarks on my system (specs below). I ran each three times. There's a bit of variance here, just from eye-balling it.

Before(1):

```
bounded0_select_both      Rust crossbeam-channel   6.365 sec
bounded0_select_rx        Rust crossbeam-channel   1.394 sec
bounded1_select_both      Rust crossbeam-channel   0.921 sec
bounded1_select_rx        Rust crossbeam-channel   0.800 sec
bounded_select_both       Rust crossbeam-channel   0.382 sec
bounded_select_rx         Rust crossbeam-channel   0.391 sec
unbounded_select_both     Rust crossbeam-channel   0.350 sec
unbounded_select_rx       Rust crossbeam-channel   0.371 sec
```

Before(2):

```
bounded0_select_both      Rust crossbeam-channel   6.146 sec
bounded0_select_rx        Rust crossbeam-channel   1.359 sec
bounded1_select_both      Rust crossbeam-channel   0.990 sec
bounded1_select_rx        Rust crossbeam-channel   0.760 sec
bounded_select_both       Rust crossbeam-channel   0.392 sec
bounded_select_rx         Rust crossbeam-channel   0.390 sec
unbounded_select_both     Rust crossbeam-channel   0.370 sec
unbounded_select_rx       Rust crossbeam-channel   0.367 sec
```

Before(3):

```
bounded0_select_both      Rust crossbeam-channel   6.193 sec
bounded0_select_rx        Rust crossbeam-channel   1.382 sec
bounded1_select_both      Rust crossbeam-channel   0.972 sec
bounded1_select_rx        Rust crossbeam-channel   0.783 sec
bounded_select_both       Rust crossbeam-channel   0.393 sec
bounded_select_rx         Rust crossbeam-channel   0.390 sec
unbounded_select_both     Rust crossbeam-channel   0.361 sec
unbounded_select_rx       Rust crossbeam-channel   0.371 sec
```

After(1):

```
bounded0_select_both      Rust crossbeam-channel   5.410 sec
bounded0_select_rx        Rust crossbeam-channel   1.320 sec
bounded1_select_both      Rust crossbeam-channel   1.032 sec
bounded1_select_rx        Rust crossbeam-channel   0.706 sec
bounded_select_both       Rust crossbeam-channel   0.367 sec
bounded_select_rx         Rust crossbeam-channel   0.358 sec
unbounded_select_both     Rust crossbeam-channel   0.368 sec
unbounded_select_rx       Rust crossbeam-channel   0.337 sec
```

After(2):

```
bounded0_select_both      Rust crossbeam-channel   5.409 sec
bounded0_select_rx        Rust crossbeam-channel   1.360 sec
bounded1_select_both      Rust crossbeam-channel   0.837 sec
bounded1_select_rx        Rust crossbeam-channel   0.716 sec
bounded_select_both       Rust crossbeam-channel   0.369 sec
bounded_select_rx         Rust crossbeam-channel   0.355 sec
unbounded_select_both     Rust crossbeam-channel   0.349 sec
unbounded_select_rx       Rust crossbeam-channel   0.339 sec
```

After(3):

```
bounded0_select_both      Rust crossbeam-channel   6.596 sec
bounded0_select_rx        Rust crossbeam-channel   1.365 sec
bounded1_select_both      Rust crossbeam-channel   0.869 sec
bounded1_select_rx        Rust crossbeam-channel   0.728 sec
bounded_select_both       Rust crossbeam-channel   0.367 sec
bounded_select_rx         Rust crossbeam-channel   0.366 sec
unbounded_select_both     Rust crossbeam-channel   0.335 sec
unbounded_select_rx       Rust crossbeam-channel   0.341 sec
```

It kind of seems like a wash to me?

I specifically ran the above on one of my idle machines too. Here is the CPU info:

```
$ lscpu
Architecture:        x86_64
CPU op-mode(s):      32-bit, 64-bit
Byte Order:          Little Endian
Address sizes:       39 bits physical, 48 bits virtual
CPU(s):              4
On-line CPU(s) list: 0-3
Thread(s) per core:  1
Core(s) per socket:  4
Socket(s):           1
NUMA node(s):        1
Vendor ID:           GenuineIntel
CPU family:          6
Model:               158
Model name:          Intel(R) Core(TM) i5-7600 CPU @ 3.50GHz
Stepping:            9
CPU MHz:             800.327
CPU max MHz:         4100.0000
CPU min MHz:         800.0000
BogoMIPS:            7010.00
Virtualization:      VT-x
L1d cache:           32K
L1i cache:           32K
L2 cache:            256K
L3 cache:            6144K
NUMA node0 CPU(s):   0-3
Flags:               fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf tsc_known_freq pni pclmulqdq dtes64 monitor ds_cpl vmx smx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid sse4_1 sse4_2 x2apic movbe popcnt aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault epb invpcid_single pti tpr_shadow vnmi flexpriority ept vpid ept_ad fsgsbase tsc_adjust bmi1 hle avx2 smep bmi2 erms invpcid rtm mpx rdseed adx smap clflushopt intel_pt xsaveopt xsavec xgetbv1 xsaves dtherm ida arat pln pts hwp hwp_notify hwp_act_window hwp_epp
```